Fix problem with deleting work executable
From Andrew 2018-05-18

The worker programs were not shutting down properly because of their connection to the master. There is an MPI function that frees this connection and then they close down properly and the worker exe file can be deleted / replaced.

The MPI function needed is called Comm_free. I have added it to the interface module MPI and the library module MPImslib and MPImsimp. The new function is then used in module MPIworker in the Close procedure which gets called at the end of Init in ParallelWorker.