Distribute using MPI_Comm_spawn rather than ports
Adds Comm_spawn procedure to MPI interface, and uses this to spawn and connect the master and the workers.

This replaces the old system in which workers the workers are set up separately and then linked to the master via ports.

This new method is more portable, since it avoids using Dialog.RunExternal, which is not available on Linux yet.

Patches from Andrew 2018-06-21 09:06 and 2018-06-21 09:08